Get paths to example files

The envlogger-package comes bundled with several sample files in its inst/extdata directory. env_example() provides a quick shortcut to access the paths to these sample files.

env_example(pattern = ".", just_dir = FALSE, delete_new_metadata_files = FALSE)

Arguments

pattern

Patterns to select one or more example files. pattern is vectorized, so all values supplied are used. If NULL (default) all example files are returned.

just_dir

If TRUE only folder paths are returned.

delete_new_metadata_files

When executing the examples provided for some functions in the envlogger-package, new metadata files may be created, which in turn may limit the scope of other examples. If set to TRUE, any metadata file that isn't part of the original example files bundle is deleted, allowing for a clean slate.

Value

A named list with the full path to one or more example files, grouped by EnvLogger file type. If just_dir = TRUE, a character vector with folder paths.

Details

The sample files provided consist of EnvLogger reports, logfiles and metadata files concerning several years of data from a few rocky shores in northern Portugal and Norway. The content in some of the files has been modified to introduce issues that exemplify some of the package's quality control features. Files include:

"normal" data, just temperature - nozzz, ptzzw, and ptzzy

"normal" data, just temperature, different format - notenv

"normal" data, temperature and pressure - press

"normal" data, temperature and humidity - humid

"abnormal" data, with several quality issues - issue

See also

Examples

# Get the file paths of all example files
env_example()
#> $log
#> [1] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/nozzz/2024-06-05/log_2024-06-05_063724.csv"        
#> [2] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/press/2025-01-31/log_2025-01-31_093212.csv"        
#> [3] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzw/2024-05-25/log_2024-05-25_095701.csv"        
#> [4] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2024-01-12/log_2024-01-12_082832_bad_col.csv"
#> [5] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14/log_2025-01-14_070601.csv"        
#> 
#> $metadata
#> [1] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/issue/2024-01-16/improper_file_name_meta.csv"                          
#> [2] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/issue/2024-01-16/issuemh04a_0434_ef00_4b29_0a-20240116_132500_meta.csv"
#> [3] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/notenv/2016-10-19/4c0000002418fb41_meta.csv"                           
#> 
#> $report
#>  [1] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/humid/2025-05-06/humidsc01a-0482_ac00_f27d_01-20250506_095113.csv"        
#>  [2] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/issue/2024-01-16/improper_file_name.csv"                                  
#>  [3] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/issue/2024-01-16/issuemh04a_0434_ef00_4b29_0a-20240116_132500.csv"        
#>  [4] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/issue/2024-05-25/issuemc01a-04ad_e500_302d_06-20240525_101116_no_data.csv"
#>  [5] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/issue/2024-05-25/issuemc02a-04b4_3e00_c90f_0e-20240525_101121_low.csv"    
#>  [6] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/issue/2024-05-25/issuemh01a-04d3_5e00_b818_02-20240525_095727_gap.csv"    
#>  [7] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/issue/2024-05-25/issuemh02a-0449_0c00_2310_05-20240525_095804_1970.csv"   
#>  [8] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/notenv/2016-10-19/4c0000002418fb41.txt"                                   
#>  [9] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/nozzz/2024-06-05/nozzzmc01a-0425_e500_051e_0b-20240605_063737.csv"        
#> [10] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/nozzz/2024-06-05/nozzzmc02a-04da_2800_d33a_0d-20240605_063925.csv"        
#> [11] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/nozzz/2024-06-05/nozzzmh01a-04fd_0b00_3805_0e-20240605_064147.csv"        
#> [12] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/nozzz/2024-06-05/nozzzmh02a-04a5_bb00_521c_0f-20240605_064246.csv"        
#> [13] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/press/2025-01-31/pressmc01b-0405_f000_5411_01-20250131_093703.csv"        
#> [14] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzw/2024-05-25/ptzzwmc01a-04ad_e500_302d_06-20240525_101015.csv"        
#> [15] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzw/2024-05-25/ptzzwmc02a-04b4_3e00_c90f_0e-20240525_101121.csv"        
#> [16] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzw/2024-05-25/ptzzwmh01a-04d3_5e00_b818_02-20240525_095727.csv"        
#> [17] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzw/2024-05-25/ptzzwmh02a-0449_0c00_2310_05-20240525_095804.csv"        
#> [18] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2024-01-12/ptzzymc01a-04ce_ce00_4114_06-20240112_084047.csv"        
#> [19] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2024-01-12/ptzzymc02a-04cb_8f00_e758_05-20240112_083906.csv"        
#> [20] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2024-01-12/ptzzymh01a-04dc_6700_dd59_0f-20240112_082904.csv"        
#> [21] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2024-01-12/ptzzymh02a-04cb_cc00_1507_0c-20240112_083030.csv"        
#> [22] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14/ptzzymc01a-04ce_ce00_4114_06-20250114_111522.csv"        
#> [23] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14/ptzzymc02a-14cb_8f00_e758_05-20250114_111020.csv"        
#> [24] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14/ptzzymh01a-04dc_6700_dd59_0f-20250114_111804.csv"        
#> [25] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14/ptzzymh02a-04cb_cc00_1507_0c-20250114_111701.csv"        
#> 
#> $unsupported
#> [1] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/issue/2024-05-25/unsupported_file_1.csv"
#> [2] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/issue/2024-05-25/unsupported_file_2.csv"
#> [3] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/issue/2024-05-25/unsupported_file_3.txt"
#> [4] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/notenv/2016-10-19/670000002f037841.txt" 
#> 

# Get the file paths to a target serial
env_example("04CB_CC00_1507_0C")
#> $report
#> [1] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2024-01-12/ptzzymh02a-04cb_cc00_1507_0c-20240112_083030.csv"
#> [2] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14/ptzzymh02a-04cb_cc00_1507_0c-20250114_111701.csv"
#> 

# Get the file paths to any example files matching a search string
env_example("20250114")
#> $report
#> [1] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14/ptzzymc01a-04ce_ce00_4114_06-20250114_111522.csv"
#> [2] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14/ptzzymc02a-14cb_8f00_e758_05-20250114_111020.csv"
#> [3] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14/ptzzymh01a-04dc_6700_dd59_0f-20250114_111804.csv"
#> [4] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14/ptzzymh02a-04cb_cc00_1507_0c-20250114_111701.csv"
#> 

# Get only logfiles
env_example()$log
#> [1] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/nozzz/2024-06-05/log_2024-06-05_063724.csv"        
#> [2] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/press/2025-01-31/log_2025-01-31_093212.csv"        
#> [3] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzw/2024-05-25/log_2024-05-25_095701.csv"        
#> [4] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2024-01-12/log_2024-01-12_082832_bad_col.csv"
#> [5] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14/log_2025-01-14_070601.csv"        

# Get only folder paths
env_example(just_dir = TRUE)
#> [1] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/issue/2024-05-25" 
#> [2] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/notenv/2016-10-19"
#> [3] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/humid/2025-05-06" 
#> [4] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/issue/2024-01-16" 
#> [5] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/nozzz/2024-06-05" 
#> [6] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/press/2025-01-31" 
#> [7] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzw/2024-05-25" 
#> [8] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2024-01-12" 
#> [9] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14" 

# 'env_example()' is vectorized, meaning that multiple search strings can be used at once
env_example(c("ptzzy", "nozzz"))
#> $log
#> [1] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/nozzz/2024-06-05/log_2024-06-05_063724.csv"        
#> [2] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2024-01-12/log_2024-01-12_082832_bad_col.csv"
#> [3] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14/log_2025-01-14_070601.csv"        
#> 
#> $report
#>  [1] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/nozzz/2024-06-05/nozzzmc01a-0425_e500_051e_0b-20240605_063737.csv"
#>  [2] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/nozzz/2024-06-05/nozzzmc02a-04da_2800_d33a_0d-20240605_063925.csv"
#>  [3] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/nozzz/2024-06-05/nozzzmh01a-04fd_0b00_3805_0e-20240605_064147.csv"
#>  [4] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/nozzz/2024-06-05/nozzzmh02a-04a5_bb00_521c_0f-20240605_064246.csv"
#>  [5] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2024-01-12/ptzzymc01a-04ce_ce00_4114_06-20240112_084047.csv"
#>  [6] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2024-01-12/ptzzymc02a-04cb_8f00_e758_05-20240112_083906.csv"
#>  [7] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2024-01-12/ptzzymh01a-04dc_6700_dd59_0f-20240112_082904.csv"
#>  [8] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2024-01-12/ptzzymh02a-04cb_cc00_1507_0c-20240112_083030.csv"
#>  [9] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14/ptzzymc01a-04ce_ce00_4114_06-20250114_111522.csv"
#> [10] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14/ptzzymc02a-14cb_8f00_e758_05-20250114_111020.csv"
#> [11] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14/ptzzymh01a-04dc_6700_dd59_0f-20250114_111804.csv"
#> [12] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14/ptzzymh02a-04cb_cc00_1507_0c-20250114_111701.csv"
#>