List EnvLogger files

Search paths for EnvLogger files (EnvLogger reports, logfiles and metadata files).

env_ls(paths, avoid_pattern = NULL, list_unsupported = FALSE)

Arguments

paths

A character vector of one or more paths. May include folders, which are searched recursively.

avoid_pattern

A character vector of one or more patterns. File paths matching to avoid_pattern are dropped.

list_unsupported

Should unsupported files be included in the output?

Value

A tibble with nine columns, including the file paths (path), two columns to indicate the type of files encountered, and six columns to facilitate quick filtering of specific file categories.

In more detail, the two columns that indicate the type of files encountered are $int (numeric) and $chr (character), and their values are as follows:

1 and report if an EnvLogger report or a valid report with a different format and an associated metadata file

2 and log if an EnvLogger logfile

3 and metadata if an EnvLogger metadata file

0 and unsupported if none of the three

The remaining columns are logical vectors for quick filtering, namely:

rep - path points to a valid report of any kind?

rep_env - path points to an EnvLogger report?

rep_other - path points to a report with a different format?

has_met - path points to a report with an associated metadata file?

log - path points to an EnvLogger logfile?

met - path points to an EnvLogger metadata file?

Details

env_ls() determines if the paths provided point to EnvLogger reports, logfiles, metadata files or none of the three. This is done by checking for the presence of certain strings in the header of each file (such as "www.electricblue.eu, Portugal"). As a result, reports generated using the the earliest versions of the EnvLogger_Viewer app may not be recognized. Also, if the structure of a file has been changed (e.g., by opening and saving in Excel), it may no longer be recognized.

Non-EnvLogger data reports

Otherwise unsupported files that are accompanied by an associated metadata file are assumed to be valid report files created by devices outside of the EnvLogger ecosystem, and will therefore be listed as "valid reports" (but this is no guarantee that they will be successfully read by downstream functions).

See also

Examples

paths <- env_example()
env_ls(paths)
#> # A tibble: 33 × 9
#>    path                    int chr   rep   rep_env rep_other has_met log   met  
#>    <chr>                 <dbl> <chr> <lgl> <lgl>   <lgl>     <lgl>   <lgl> <lgl>
#>  1 /private/var/folders…     1 repo… TRUE  TRUE    FALSE     FALSE   FALSE FALSE
#>  2 /private/var/folders…     1 repo… TRUE  TRUE    FALSE     TRUE    FALSE FALSE
#>  3 /private/var/folders…     1 repo… TRUE  TRUE    FALSE     TRUE    FALSE FALSE
#>  4 /private/var/folders…     1 repo… TRUE  TRUE    FALSE     FALSE   FALSE FALSE
#>  5 /private/var/folders…     1 repo… TRUE  TRUE    FALSE     FALSE   FALSE FALSE
#>  6 /private/var/folders…     1 repo… TRUE  TRUE    FALSE     FALSE   FALSE FALSE
#>  7 /private/var/folders…     1 repo… TRUE  TRUE    FALSE     FALSE   FALSE FALSE
#>  8 /private/var/folders…     1 repo… TRUE  FALSE   TRUE      TRUE    FALSE FALSE
#>  9 /private/var/folders…     1 repo… TRUE  TRUE    FALSE     FALSE   FALSE FALSE
#> 10 /private/var/folders…     1 repo… TRUE  TRUE    FALSE     FALSE   FALSE FALSE
#> # ℹ 23 more rows