Parse structured logger ids

If the logger ids used follow a stable naming scheme, parse_id() can be used to generate new columns that hold each naming segment. parse_id_cctbon() provides a quick shortcut for handling data that follows the CCTBON id scheme.

parse_id(env, div = NULL, fields = NULL, standard = NULL)

parse_id_cctbon(env)

Arguments

env

The output of a call to read_env() (either directly, of the $report element).

div

A sequence of numbers defining the length of each id element. Can be provided as a number (1111123445) or as a string ("1111123445"). Numbers must be consecutive and same numbers must be contiguous. Same numbers define the position and width of the desired element in the ids.

fields

A list providing the names and data types of each id elements. fields must have as many elements as those outlined in div. Certain downstream functions such as add_info() expect a column named site coding location name. Check the Fields section below for more information on the structure of fields.

standard

In alternative to setting div and fields, provide the name of one of the supported naming schemes and the other args will be set automatically to the expected values. Current schemes supported are: 'cctbon'. Check the details sections below for more information.

Value

A tibble with EnvLogger report data, augmented with new columns resulting from the parsing of data report ids.

standard = CCTBON

If the logger ids follow scheme used by the CCTBON network (https://www.coastalwarming.com/cctbon), simply use parse_id_cctbon() instead.

To conform with the CCTBON scheme, ids must:

always be 10 chr long

always contain the following elements in the exact order as outlined next:

5 chr (#1-5) for site name (must be unique in the network)

1 chr (#6) for shore level (l, m or t = low, mid or top intertidal, s = supratidal)

1 chr (#7) for solar exposure (c = cold, h = hot)

2 chr (#8-9) for microhabitat replicate number (01, 02, 03, ...)

1 chr (#10) for logger position replicate number (a, b, c, ...)

Fields

We can use the CCTBON scheme to better illustrate the structure of fields. In that case:

div = "1111123445"

fields = list(site = "fct", lvl = c("l","m","t","s"), exp = c("c","h"), rep1 = "num", rep2 = "chr")

This will result in chars #1-5 being assigned to column site, char #6 to column lvl, char #7 to column exp, and so on. Moreover, the data types are interpreted as follows:

fct results in a factor with levels set to the unique values available

chr results in a character vector (as.character)

num results in a numeric vector (as.numeric)

when the type provided has a length greater than one (as is the case for lvl and exp in the CCTBON scheme), results in a factor with levels set to the values provided as type

As long as (a) the number of fields provided matches the number of different numbers supplied in div, (b) fields follows the structure field name = data type, (c) data types are one of 'chr', 'num', 'fct', or c('a', 'b'), and (d) data type conversions are possible without generating NAs, any file scheme can be parsed.

See also

Examples

env <- read_env(
   env_example(c("ptzzy", "ptzzw")),
   new_interval = 60,
   show_progress = FALSE,
   show_warnings = FALSE)

# If ids follow a supported id namind standard
parse_id(env, standard = "cctbon")
#> # A tibble: 8 × 17
#>   id       serial site  lvl   exp    rep1 rep2  type  v_log press hum   data    
#>   <chr>    <chr>  <fct> <fct> <fct> <dbl> <chr> <chr> <chr> <lgl> <lgl> <list>  
#> 1 ptzzwmc… 04ad … ptzzw m     c         1 a     t     7.3   FALSE FALSE <tibble>
#> 2 ptzzwmc… 04b4 … ptzzw m     c         2 a     t     7.3   FALSE FALSE <tibble>
#> 3 ptzzwmh… 04d3 … ptzzw m     h         1 a     t     7.3   FALSE FALSE <tibble>
#> 4 ptzzwmh… 0449 … ptzzw m     h         2 a     t     7.3   FALSE FALSE <tibble>
#> 5 ptzzymc… 04ce … ptzzy m     c         1 a     t     2.4   FALSE FALSE <tibble>
#> 6 ptzzymc… 04cb … ptzzy m     c         2 a     t     2.4   FALSE FALSE <tibble>
#> 7 ptzzymh… 04dc … ptzzy m     h         1 a     t     2.4   FALSE FALSE <tibble>
#> 8 ptzzymh… 04cb … ptzzy m     h         2 a     t     2.4   FALSE FALSE <tibble>
#> # ℹ 5 more variables: nrow <dbl>, min <dbl>, max <dbl>, t0 <dttm>, t1 <dttm>
parse_id_cctbon(env) # even shorter for cctbon
#> # A tibble: 8 × 17
#>   id       serial site  lvl   exp    rep1 rep2  type  v_log press hum   data    
#>   <chr>    <chr>  <fct> <fct> <fct> <dbl> <chr> <chr> <chr> <lgl> <lgl> <list>  
#> 1 ptzzwmc… 04ad … ptzzw m     c         1 a     t     7.3   FALSE FALSE <tibble>
#> 2 ptzzwmc… 04b4 … ptzzw m     c         2 a     t     7.3   FALSE FALSE <tibble>
#> 3 ptzzwmh… 04d3 … ptzzw m     h         1 a     t     7.3   FALSE FALSE <tibble>
#> 4 ptzzwmh… 0449 … ptzzw m     h         2 a     t     7.3   FALSE FALSE <tibble>
#> 5 ptzzymc… 04ce … ptzzy m     c         1 a     t     2.4   FALSE FALSE <tibble>
#> 6 ptzzymc… 04cb … ptzzy m     c         2 a     t     2.4   FALSE FALSE <tibble>
#> 7 ptzzymh… 04dc … ptzzy m     h         1 a     t     2.4   FALSE FALSE <tibble>
#> 8 ptzzymh… 04cb … ptzzy m     h         2 a     t     2.4   FALSE FALSE <tibble>
#> # ℹ 5 more variables: nrow <dbl>, min <dbl>, max <dbl>, t0 <dttm>, t1 <dttm>

# If using a different scheme
parse_id(
   env,
   div = "1111123445",
   fields = list(
       site = "fct",
       lvl  = c("l","m","t","s"),
       exp  = c("c","h"),
       rep1 = "num",
       rep2 = "chr")
       )
#> # A tibble: 8 × 17
#>   id       serial site  lvl   exp    rep1 rep2  type  v_log press hum   data    
#>   <chr>    <chr>  <fct> <fct> <fct> <dbl> <chr> <chr> <chr> <lgl> <lgl> <list>  
#> 1 ptzzwmc… 04ad … ptzzw m     c         1 a     t     7.3   FALSE FALSE <tibble>
#> 2 ptzzwmc… 04b4 … ptzzw m     c         2 a     t     7.3   FALSE FALSE <tibble>
#> 3 ptzzwmh… 04d3 … ptzzw m     h         1 a     t     7.3   FALSE FALSE <tibble>
#> 4 ptzzwmh… 0449 … ptzzw m     h         2 a     t     7.3   FALSE FALSE <tibble>
#> 5 ptzzymc… 04ce … ptzzy m     c         1 a     t     2.4   FALSE FALSE <tibble>
#> 6 ptzzymc… 04cb … ptzzy m     c         2 a     t     2.4   FALSE FALSE <tibble>
#> 7 ptzzymh… 04dc … ptzzy m     h         1 a     t     2.4   FALSE FALSE <tibble>
#> 8 ptzzymh… 04cb … ptzzy m     h         2 a     t     2.4   FALSE FALSE <tibble>
#> # ℹ 5 more variables: nrow <dbl>, min <dbl>, max <dbl>, t0 <dttm>, t1 <dttm>