Read EnvLogger data

Reads and organizes EnvLogger data and associated information. The following steps are executed sequentially:

step 1 – read files

step 2 – apply fixes

step 3 – check quality

step 4 – correct RTC drift

step 5 – interpolate to round time

step 6 – join by serial

step 7 – join by id

step 8 – keep only full days

step 9 – auto-generate metadata files

read_env(
  paths,
  avoid_pattern = NULL,
  log_summary = TRUE,
  show_progress = TRUE,
  show_warnings = TRUE,
  read_data = TRUE,
  apply_fixes = TRUE,
  qual_lims = list(min = -30, max = +70, t0 = "2000-01-01", t1 = Sys.Date()),
  correct_rtc_drift = TRUE,
  new_interval = NULL,
  join_serials = TRUE,
  join_ids = TRUE,
  join_ids_trim_d1 = TRUE,
  overlap_max_mins = 60 * 3,
  keep_flags = FALSE,
  full_days = TRUE,
  auto_generate_fixes = FALSE,
  auto_implement_fixes = FALSE
)

Arguments

paths

A character vector of one or more paths. May include folders, which are searched recursively.

avoid_pattern

A character vector of one or more patterns. File paths matching to avoid_pattern are dropped.

log_summary

Should logfile entries be provided in full (FALSE) or should they be made more compact by only including the last entry for each serial?

show_progress

Should progress messages be printed?

show_warnings

If issues are detected, should final warning messages be printed?

read_data

If set to FALSE, only the header information is gathered from the data reports. This is much quicker, and can be used to assess structure and issues from a large database before importing all data.

apply_fixes

Whether to implement the fixes found in the metadata files.

qual_lims

A list with the elements min, max, t0, t1 (names are strict). These values will be used to identify quality issues in each data report (values outside the windows provided are flagged as issues). min and max correspond to the low and high temperature thresholds. t0 and t1 correspond to the time window accepted for the timestamps. Default values are min = -30, max = +70, t0 = "2000-01-01" and t1 = Sys.Date(), i.e., today.

correct_rtc_drift

Whether to expand or compress timestamps to account for recorded RTC drift.

new_interval

Desired sampling rate in minutes. If left NULL no interpolation takes place and data remains raw. If provided, data is interpolated to every new_interval-minutes. This is important for analyses that require data from all loggers to be at the same time points (every hour, every half-hour). Interpolation is aware of eventual gaps in the data and does not generate artificial data where none exists.

join_serials

Whether to join logger reports relating to the same serial numbers.

join_ids

If set to TRUE, join_serials is forced to TRUE as well. Whether to join logger reports relating to the same ids (information specific to individual report files is discarded).

join_ids_trim_d1

When joining by ids, should the day when serials change be discarded?

overlap_max_mins

Maximum allowed overlap length in minutes. When data from two files overlaps by less than the stipulated number of minutes, the situation is normal and handled silently. Otherwise, an issue is reported.

keep_flags

Whether to retain columns flagging data processing steps, which are normally only used internally by read_env() and discarded before printing the output. Only relevant for debugging errors.

full_days

Should logger data only be kept for days in which all data points were collected? (e.g., if sampling rate is 60 minutes, full days must have 24 data points). This option is only available if new_interval is provided.

auto_generate_fixes

If issues are identified for which a simple fix can be automatically determined, should the corresponding metadata files be automatically created?

auto_implement_fixes

If set to TRUE, auto_generate_fixes is forced to TRUE as well. After automatically generating metadata files to address issues identified, should data be re-imported again to immediately implement those fixes?

Value

A list with EnvLogger information (and data) from all reports, logfiles and metadata found in the target paths, divided into $log, $metadata, and $report. Additionally, if issues are found, the returned list may also include $issues, $files_with_issues, and $files_created.

RTC drift correction

The correction implemented is linear and assumes that RTC drift accumulates at the same rate throughout the deployment. This assumption comes from the fact that all EnvLoggers feature temperature-compensated RTCs, which should result in drift being independent of temperature and only be affected by deployment length. EnvLoggers T2.4 accumulate about 10 to 15 minutes of RTC drift over a one-year deployment, which may be important to correct. EnvLoggers T7.3 and beyond feature a much more precise RTC and accumulate only 1 to 2 minutes of drift every year. The correction applied is based on tdiff, a numeric value in seconds that represents the difference between logger RTC time (obs; observed) and smartphone time (ref; reference) as (obs - ref). This means that if the logger RTC (obs) ticks quicker than ref, tdiff becomes positive.

drift = logger - smartphone

pos -> logger time was faster -> timestamps must be compressed

neg -> logger time was slower -> timestamps must be expanded

Interpolation to round times

Perform linear interpolation to tidy EnvLogger data so that there's one reading every hour (00 mins, 00 secs) or other specified interval; !ALWAYS! applied before joining data from different reports, as otherwise the interpolation would generate artificial data filling any gap that may be present.

See also

Examples

# raw data from a single EnvLogger data report (data as is)
paths <- env_example("ptzzy")$report[[1]]
read_env(paths, correct_rtc_drift = FALSE, full_days = FALSE)
#> ℹ found 1 file in 1 folder
#> ✔ files read (n = 1) [1 reports]
#> ✔ no fix details found
#> ✔ quality checks passed by all reports
#> ✔ no reports to join by serial
#> ✔ no reports to join by id
#> $report
#> # A tibble: 1 × 12
#>   id         serial           type  v_log press hum   data      nrow   min   max
#>   <chr>      <chr>            <chr> <chr> <lgl> <lgl> <list>   <int> <dbl> <dbl>
#> 1 ptzzymc01a 04ce ce00 4114 … t     2.4   FALSE FALSE <tibble>  6873   7.6  24.9
#> # ℹ 2 more variables: t0 <dttm>, t1 <dttm>
#> 

# folder with multiple data files from a single site
paths <- env_example("ptzzy", just_dir = TRUE)
read_env(paths, new_interval = 60)
#> ℹ found 10 files in 2 folders
#> ✔ files read (n = 10) [8 reports, 2 logs]
#> ✔ no fix details found
#> ✔ quality checks passed by all reports
#> ✔ RTC clock drifts corrected
#> ✔ data interpolated to every 60 mins
#> ✔ reports joined by serial (n = 7 -> 5)
#> ✔ reports joined by id ... (n = 5 -> 4)
#> ✔ data filtered to only keep full days
#> Warning: ℹ issues detected (total n = 2)
#> ℹ type ?create_metadata_file() to learn more about the issues listed below
#> ℹ access paths to files with unresolved issues using $files_with_issues
#> ℹ access paths to newly created metadata files using $files_created
#>   
#> ✖ issues that remain unresolved (total n = 2)
#> → files that fully overlap other files have been ignored
#> → partial overlaps were resolved automatically, but consider doublechecking
#> ptzzymh02a-04cb_cc00_1507_0c-20240112_083030.csv .. (full overlap)
#> ptzzymc02a-14cb_8f00_e758_05-20250114_111020.csv .. (partial overlap)
#>   
#> dropped unsupported columns for the following logfiles:
#> log_2024-01-12_082832_bad_col.csv
#> $log
#> # A tibble: 6 × 8
#>   time                id         serial            v_log   int   res pass  dev  
#>   <dttm>              <chr>      <chr>             <dbl> <dbl> <dbl> <chr> <chr>
#> 1 2025-01-14 11:07:45 ptzzymh03a 0427 4000 911e 02   2.4  3600   0.1 1111  armo…
#> 2 2025-01-14 11:10:24 ptzzymc02a 04cb 8f00 e758 05   2.4  3600   0.1 1111  armo…
#> 3 2025-01-14 11:15:26 ptzzymc01a 04ce ce00 4114 06   2.4  3600   0.1 1111  armo…
#> 4 2025-01-14 11:16:10 ptzzymc03a 04a5 8f00 c502 10   2.4  3600   0.1 1111  armo…
#> 5 2025-01-14 11:17:05 ptzzymh02a 04cb cc00 1507 0c   2.4  3600   0.1 1111  armo…
#> 6 2025-01-14 11:18:08 ptzzymh01a 04dc 6700 dd59 0f   2.4  3600   0.1 1111  armo…
#> 
#> $report
#> # A tibble: 4 × 12
#>   id         serial           type  v_log press hum   data      nrow   min   max
#>   <chr>      <chr>            <chr> <chr> <lgl> <lgl> <list>   <dbl> <dbl> <dbl>
#> 1 ptzzymc01a 04ce ce00 4114 … t     2.4   FALSE FALSE <tibble> 12432  7.95  27.1
#> 2 ptzzymc02a 04cb 8f00 e758 … t     2.4   FALSE FALSE <tibble> 12408 10.0   22.8
#> 3 ptzzymh01a 04dc 6700 dd59 … t     2.4   FALSE FALSE <tibble> 12408  6.48  37.2
#> 4 ptzzymh02a 04cb cc00 1507 … t     2.4   FALSE FALSE <tibble> 12432  5.6   39.7
#> # ℹ 2 more variables: t0 <dttm>, t1 <dttm>
#> 
#> $issues
#> # A tibble: 2 × 6
#>   step        path_data                       path_meta issue new_vals     fixed
#>   <chr>       <chr>                           <chr>     <chr> <list>       <lgl>
#> 1 join_serial /private/var/folders/_3/fv_x_0… ""        full… <NULL>       FALSE
#> 2 join_id     /private/var/folders/_3/fv_x_0… "/privat… part… <named list> FALSE
#> 
#> $files_with_issues
#> [1] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2024-01-12/ptzzymh02a-04cb_cc00_1507_0c-20240112_083030.csv"
#> [2] "/private/var/folders/_3/fv_x_0wd36v0ddvndjlqjwfw0000gn/t/rtmpnbd0el/temp_libpath23822b2de41d/envlogger/extdata/ptzzy/2025-01-14/ptzzymc02a-14cb_8f00_e758_05-20250114_111020.csv"
#>