Group EnvLogger data by id elements and summarise it

Takes an EnvLogger tibble resulting from a call to read_env(), groups data by grouping variables and summarises it. In the simplest case, grouping is only done along time (by_day). However, if logger ids have been broken into id elements with parse_id() or parse_id_cctbon(), grouping can also be done across id elements. For example, if data follows the CCTBON naming scheme, data can be grouped by site, lvl (shore height level) or exp (exposure to solar radiation). This function exists mainly to condition data for plot_env(), but can also be useful by itself.

Note that summarise_env can be called on the output of a previous call to summarise_env, allowing, for example, for first computing daily maximums, and then on a second call, to request an average of those daily maximums across all loggers sharing the same microhabitat.

summarise_env(
  env,
  var = "temp",
  fun_list = NULL,
  by_day = TRUE,
  by_site = FALSE,
  by_group = FALSE,
  site = "site",
  group = c("lvl", "exp"),
  same_days = FALSE,
  roll_days = 0,
  by_day_lim = 50
)

Arguments

env

The output of a call to read_env() (either directly, of the $report element). Additionally, a call to parse_id() or add_info() ensures more grouping options. In alternative, the output of a call to summarise_env.

var

The name of a single column over which the function(s) in fun_list will be applied. Defaults to temp. Can also be set to hum or press if wanting to summarise using humidity or pressure data, respectively. Can also be any other name, as long as it exists as colnames and codes for a data type that will result in a single numeric value when processed by all functions provided in fun_list.

fun_list

Named list with one or more functions to be applied to var. Each function must return a single numeric value. The names of the columns derived from applying the functions provided are taken from the names of fun_list. If fun has no names, generic names are generated. If not provided, the following stats are computed: min, q10 (10th quantile), q25, average, q50, q75, q90 and max.

by_day

Whether to summarise values daily or at the native sampling interval.

by_site

Whether to group by site.

by_group

Whether to group by group.

site

The name of the column that codes for site. If a vector with more than one element, elements are combined to form a new column by pasting values together. Only relevant if by_site = TRUE. Can be omitted if data follows the CCTBON naming scheme.

group

One or more column names to combine as a grouping factor. If a vector with more than one element, elements are combined to form a new column by pasting values together. Only relevant if by_group = TRUE. Can be omitted if data follows the CCTBON naming scheme (in which case lvl and exp are used).

same_days

Should the output only include data for days for which all distinct grouping ids (site + group) have data?

roll_days

Compute a rolling mean over the x days before each data point. If 0 (default), no rolling mean is applied.

by_day_lim

Maximum number (in thousands) of collective rows accepted when by_day = FALSE. If the aggregate number of rows exceeds this value, by_day will be automatically set to TRUE. This is to prevent an unintentional call requesting the computation of summary stats over datasets that are too large, a situation that may lead to R stalling.

Value

A tibble with unnested data containing all the combinations of id and t (time), as well as site and group if by_site = TRUE and by_group = TRUE, respectively, along with as many data summary columns as the number of functions provided using fun_list.

See also

Examples

env <- read_env(
   env_example(c("ptzzy", "ptzzw")),
   new_interval = 60,
   show_progress = FALSE,
   show_warnings = FALSE)

env <- parse_id_cctbon(env)

# daily max for each logger
daily_max <- summarise_env(
   env,
   var = "temp",
   fun_list = list(max = max),
   by_day = TRUE
   )
daily_max
#> # A tibble: 3,635 × 9
#>    id         site  lvl   exp    rep1 rep2  sgmnt t                     max
#>    <chr>      <fct> <fct> <fct> <dbl> <chr> <chr> <dttm>              <dbl>
#>  1 ptzzymc01a ptzzy m     c         1 a     1a    2023-04-01 00:00:00  16.3
#>  2 ptzzymc02a ptzzy m     c         2 a     1a    2023-04-01 00:00:00  16.4
#>  3 ptzzymh01a ptzzy m     h         1 a     1a    2023-04-01 00:00:00  17.7
#>  4 ptzzymh02a ptzzy m     h         2 a     1a    2023-04-01 00:00:00  17.0
#>  5 ptzzymc01a ptzzy m     c         1 a     1a    2023-04-02 00:00:00  16.1
#>  6 ptzzymc02a ptzzy m     c         2 a     1a    2023-04-02 00:00:00  16.1
#>  7 ptzzymh01a ptzzy m     h         1 a     1a    2023-04-02 00:00:00  19.6
#>  8 ptzzymh02a ptzzy m     h         2 a     1a    2023-04-02 00:00:00  19.2
#>  9 ptzzymc01a ptzzy m     c         1 a     1a    2023-04-03 00:00:00  15.7
#> 10 ptzzymc02a ptzzy m     c         2 a     1a    2023-04-03 00:00:00  15.7
#> # ℹ 3,625 more rows

# daily minimum, average and maximum daily_max for each site
summarise_env(
   daily_max,
   var = "max",
   fun_list = list(
       dmax_min = min,
       dmax_avg = mean,
       dmax_max = max,
       sd = sd
       ),
   by_day = TRUE,
   by_site = TRUE,
   same_days = TRUE
   )
#> # A tibble: 782 × 7
#>    id    site  t                   dmax_min dmax_avg dmax_max    sd
#>    <fct> <fct> <dttm>                 <dbl>    <dbl>    <dbl> <dbl>
#>  1 ptzzw ptzzw 2023-04-30 00:00:00     23.2     28.4     34.3  6.03
#>  2 ptzzy ptzzy 2023-04-30 00:00:00     17.1     18.8     20.6  1.56
#>  3 ptzzw ptzzw 2023-05-01 00:00:00     23.2     28.1     32.6  4.97
#>  4 ptzzy ptzzy 2023-05-01 00:00:00     16.6     18.5     19.7  1.41
#>  5 ptzzw ptzzw 2023-05-02 00:00:00     22.8     25.2     27.5  2.60
#>  6 ptzzy ptzzy 2023-05-02 00:00:00     16.4     19.3     21.6  2.68
#>  7 ptzzw ptzzw 2023-05-03 00:00:00     23.4     26.6     28.9  2.62
#>  8 ptzzy ptzzy 2023-05-03 00:00:00     16.5     18.8     21.6  2.67
#>  9 ptzzw ptzzw 2023-05-04 00:00:00     22.8     26.4     29.0  2.92
#> 10 ptzzy ptzzy 2023-05-04 00:00:00     17.2     19.7     22.3  2.66
#> # ℹ 772 more rows